Taxi service. Adding spring file
