+++ pom.xml
@@ -15,8 +15,8 @@
         <maven.checkstyle.plugin.version>3.1.1</maven.checkstyle.plugin.version>
         <maven.checkstyle.plugin.configLocation>
             checkstyle.xml
-            <!--https://raw.githubusercontent.com/mate-academy/style-guides/master/java/checkstyle.xml-->
         </maven.checkstyle.plugin.configLocation>
+        <spring.version>5.2.21.RELEASE</spring.version>
     </properties>
 
     <dependencies>
@@ -25,6 +25,16 @@
                 <artifactId>mysql-connector-j</artifactId>
                 <version>8.0.33</version>
             </dependency>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-context</artifactId>
+            <version>${spring.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-orm</artifactId>
+            <version>${spring.version}</version>
+        </dependency>
     </dependencies>
 
     <build>
