+++ src/main/java/com/taxi/config/AppConfig.java
@@ -1,0 +1,17 @@
+package com.taxi.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.core.env.Environment;
+
+@Configuration
+@PropertySource("classpath:db.properties")
+@ComponentScan(basePackages = {"com.taxi"})
+public class AppConfig {
+    private final Environment environment;
+
+    public AppConfig(Environment environment) {
+        this.environment = environment;
+    }
+}
