+++ src/main/java/com/taxi/dao/interfaces/ManufacturerDao.java
@@ -1,0 +1,17 @@
+package com.taxi.dao.interfaces;
+
+import java.util.List;
+import java.util.Optional;
+import com.taxi.model.Manufacturer;
+
+public interface ManufacturerDao {
+    Manufacturer create(Manufacturer manufacturer);
+
+    Optional<Manufacturer> get(Long id);
+
+    List<Manufacturer> getAll();
+
+    Manufacturer update(Manufacturer manufacturer);
+
+    boolean delete(Long id);
+}
