+++ src/main/java/com/taxi/exception/DataProcessingException.java
@@ -1,0 +1,7 @@
+package com.taxi.exception;
+
+public class DataProcessingException extends RuntimeException {
+    public DataProcessingException(String message, Throwable ex) {
+        super(message, ex);
+    }
+}
