+++ src/main/java/com/taxi/lib/Dao.java
@@ -1,0 +1,8 @@
+package com.taxi.lib;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Dao {
+}
