--- src/main/java/mate/jdbc/dao/interfaces/ManufacturerDao.java
@@ -1,17 +1,0 @@
-package mate.jdbc.dao.interfaces;
-
-import java.util.List;
-import java.util.Optional;
-import mate.jdbc.model.Manufacturer;
-
-public interface ManufacturerDao {
-    Manufacturer create(Manufacturer manufacturer);
-
-    Optional<Manufacturer> get(Long id);
-
-    List<Manufacturer> getAll();
-
-    Manufacturer update(Manufacturer manufacturer);
-
-    boolean delete(Long id);
-}
