--- src/main/java/mate/jdbc/exception/DataProcessingException.java
@@ -1,7 +1,0 @@
-package mate.jdbc.exception;
-
-public class DataProcessingException extends RuntimeException {
-    public DataProcessingException(String message, Throwable ex) {
-        super(message, ex);
-    }
-}
