--- src/main/java/mate/jdbc/lib/Dao.java
@@ -1,8 +1,0 @@
-package mate.jdbc.lib;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Dao {
-}
